requirements.txt
flask
psycopg2-binary
gunicorn
flask-cors
pytest
